+++ src/Pages/Management/CarTableSection.tsx
@@ -1,4 +1,4 @@
-import { CarTable } from "@/Components/CarTable";
+import { CarTable } from "@/Components/ManageComponents/CarTable";
 
 export function CarTableSection() {
     return (
